refactor(svelte2tsx): extract script re-parsing into script/parse.rs
Pure move: no logic changes, only `use` additions and `pub(super)` visibility
adjustments. The corresponding unit tests move with the code.
